Interactions

MembraneAnalysis.count_interactions — Function

count_interactions(;
    pdb_file,
    traj_file,
    fs_file,
    output_file,
    lipids,
    donors=[(lipid, lipid.head_atom) for lipid in lipids],
    acceptors=[(lipid, "PO4") for lipid in lipids],
    cutoff=6.0)

Counting the number of interactions (such as H-bonds) defined as a donor lipid atom being closer than the cutoff to a acceptor lipid atom. Saves the results in an HDF5 file with the label tag.

Keyword arguments

pdb_file: PDB structure file;

traj_files: a list of trajectory files;

fs_files: a list of corresponding HDF5 fluctuation spectrum files of the trajectory files;

output_file: output directory;

lipids: a list of lipids of type Lipid as defined in lipids.jl;

donors: list of tuples of lipids (from type Lipid) and atoms to be considered as interaction donors;

acceptors: list of tuples of lipids (from type Lipid) and atoms to be considered as interaction acceptors;

cutoff: interaction distance cutoff.