Lipids

New lipid species can be defined using the Lipid struct.

MembraneAnalysis.Lipid — Type

Lipid(
    name,
    head_atom,
    ref_atom,
    tail_atom,
    n_atoms)

A struct describing lipid species present in the system.

Fields

name: lipid name;

head_atom: the head atom of the lipid (has to be the first atom in the PDB file);

ref_atom: default reference atom for determining the position of the lipid;

tail_atom: a tail atom of the lipid (does not have to be the last atom in the PDB file);

n_atoms: number of atoms in the lipid.